+++ src/Console/Application.php
@@ -44,6 +44,7 @@ class Application extends BaseApplication
         $commands[] = new \SugarCli\Console\Command\MetadataDumpCommand();
         $commands[] = new \SugarCli\Console\Command\MetadataLoadCommand();
         $commands[] = new \SugarCli\Console\Command\MetadataStatusCommand();
+        $commands[] = new \SugarCli\Console\Command\InventoryFacterCommand();
         return $commands;
     }
 
@@ -69,4 +70,3 @@ class Application extends BaseApplication
         return parent::run(null, $output);
     }
 }
-