+++ src/Inventory/FactsProvider.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace SugarCli\Inventory;
+
+interface FactsProvider
+{
+    public function getFacts();
+}