+++ src/Inventory/FactsProvider/Hostname.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace SugarCli\Inventory\FactsProvider;
+
+use SugarCli\Inventory\FactsProvider;
+
+class Hostname implements FactsProvider
+{
+    public function getFacts()
+    {
+        return array('hostname' => gethostname());
+    }
+}