+++ src/Inventory/FactsProvider/PuppetFacter.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace SugarCli\Inventory\FactsProvider;
+
+use SugarCli\Inventory\CommandFactsProvider;
+
+class PuppetFacter extends CommandFactsProvider
+{
+    protected $cmd = 'facter --json';
+    protected $as_json = true;
+}